Backend/requirements.txt
alembic==1.16.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.2.1
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.5
declxml==0.9.1
decorator==5.1.1
defusedxml==0.7.1
et_xmlfile==2.0.0
executing==2.1.0
fastapi==0.116.1
fastjsonschema==2.20.0
filelock==3.19.1
fonttools==4.56.0
fqdn==1.5.1
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.9
imbalanced-learn==0.13.0
imblearn==0.0
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
Mako==1.3.10
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
mpmath==1.3.0
narwhals==1.41.1
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.2.6
opencv-python==4.12.0.88
openpyxl==3.1.5
optuna==4.3.0
overrides==7.7.0
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pillow==12.0.0
platformdirs==4.3.3
polars==1.33.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==6.31.1
psutil==6.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pygame==2.6.1
Pygments==2.18.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.2
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==74.1.2
six==1.16.0
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.47.3
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
toml==0.10.2
torch==2.8.0
torchvision==0.23.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20240906
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2024.1
ultralytics==8.3.235
ultralytics-thop==2.0.18
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.35.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
xgboost==3.0.2
xlrd==2.0.1
